+++ packages/fiber/src/native.tsx
@@ -1,5 +1,3 @@
-// for type declarations
-// for type declarations
 export * from './three-types'
 import * as ReactThreeFiber from './three-types'
 export { ReactThreeFiber }
@@ -15,7 +13,5 @@ export type {
   RootState,
 } from './core/store'
 export type { ThreeEvent, Events, EventManager } from './core/events'
-
-export type { Props } from './reactnative/Canvas'
-
-export * from './reactnative'
+export type { Props } from './native/Canvas'
+export * from './native'